Unify the sumcheck and MLE-check round drivers
Six entry points drove the same round loop. The sumcheck and MLE-check
variants of `prove_single`, `batch_prove`, and the write-evals wrapper
were near-identical, differing in one line: which coefficient of the
round polynomial is left off the wire.

Sumcheck drops the highest-degree coefficient; MLE-check drops the
constant term. The verifier reconstructs the dropped one from the round
claim, so the two are not interchangeable — sending the wrong one
produces a proof that fails verification.

That choice is now a small `RoundProofKind` enum consulted in exactly
one place, and the six entry points delegate to one `single` and one
`batch` driver. Public signatures are unchanged.
