+++ crates/ip-prover/src/sumcheck/mod.rs
@@ -5,6 +5,7 @@ pub mod batch;
 pub mod bivariate_product_evaluator;
 pub mod bivariate_product_mle;
 pub mod common;
+mod drive;
 pub mod gruen32;
 pub mod mle_store;
 mod mle_to_sumcheck;
